+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-07 18:01:10
+
+### Changed Files
+- `src/features/admin/components/AdminControlCenter.tsx`
+
+### Statistics
+```
+ 1 file changed, 10 insertions(+), 5 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-07 17:19:39
 
 ### Changed Files